fix 달력과 일일목표칼로리를 권장비율로 나타내주는 화면 수정
## 🏷 작업 분류

- [ ]  기능 추가
- [ ]  리팩토링
- [x]  버그 수정
- [ ]  환경 수정 및 기타

## 💡 작업 개요
달력 눌렀을때 목표칼로리의 권장비율이 나오지 않던것을 수정하였습니다!
## ***📜** 리뷰 요청사항*
